expose the test bench as a global variable
The previous solution was to inject the test bench in the tests before
they were run by using modified TestRunner and TestResult. That worked
but it was a bit complex.

The alternative is to expose the test bench as global variable of the
bench submodule. The main is now in charge of setup it up, and tests can
access it through an import:

    from squidsa.bench import bench