bench: try to use `bench.cfg` for bench configuration
Instead of expecting the bench configuration to be under conf/env/,
squidsa will now try to load `bench.cfg`, or the configuration file
passed to it with flags `-B` or `--bench`.

The reason is that boards and switches configurations are static and can
be shared with other users, whereas each user has a unique test bench
configuration.