parser/switch: use the right string to find the config file
A previous commit changed the way switches are named. If a switch
defined like this in the board was file was foud

  [switch0]
  name = "wag200g"

It would create a switch named wag200g#0. Instead of using the "name"
key directly as a key to locate the config file, it used the generated
name, which resulted in an error.

The switch parser is now updated to take both its full name and the
model as parameter, so that the right file can be found.

(Fixes: f48086a10 "parser/board: Name switch more explicitely")