resources: try to locate resources at multiple locations
Resource files were search in a mutually exclusive way: either in the
out-of-tree configuration directory that was set on the command line, or
in the `conf/` directory.

So if a user wanted to use a slightly modified board file out-of-tree
but reuse the same switch files, the whole conf/ directory had to be
copied.

Squidsa is now a bit smarter and will try to locate files in the
configuration directory passed by the user, and default to its own
directory if not found.