bench/link: fix access to wrong members
Link's members `host` and `sut` were renamed with `host_if` and `sut_if`
to make it clearer that these members were refering to interfaces, and
not to machines, but the __repr__ function hasn't been updated.

Also fix a typo. "is" was missing in the sentence "This is just a
convenient wrapper".