SSHControl: let it search its extra args by itself
For now, passing arguments to SSHControl is a bit convoluted. They must
first be parsed by the BenchParser, and retrieved by the Bench class to
be passed to SSHControl's init function.

To simplify that, and make children of the Control class all use the
same prototype for their init functions, SSHControl now takes the
address and the bench_parser as argument. So if it needs extra variables
for its configuration, it will look them up by itself.