bench configuration: accept "control" key
Bench configuration files now accept a "control" key in their host and
target section. It's expected to look like "ssh://<ip>". If such key is
not found, it will default to "local" (ie. it will execute commands on
the local machine).

For now, control subclasses still search the configuration file if they
need extra information, but ideally in the end, it will be encoded in
the URI directly.