parser/bench: store links in an ordered dictionary
BenchParser parses the bench configuration file and stores the link it
can find in a dictionary, with a mapping "LinkName -> Instance of
LinkParser". The Bench class then iterates over that dictionary to
create its own list of links.

But the dictionary is not ordered, so multiple runs might return links
in a different order. That makes dsatest unstable and less
deterministic. Replace the dictionary by an OrderedDict, which remember
the order that items were inserted into it.

Fixes #11