bench: make `--test` cumulative
This changes the behaviour of the `--test` option. It used to support
either "all", a filename (ie. a parameter ending with .py), or a
pattern. When a pattern was passed, any test cases containing the
pattern in their methods was added to the list of tests to be run.

Now, only the latter case is supported, with a slight change though:
only tests cases whose test methodes start with test_<arg> are returned.
So for instance, with --test ping, only tests with a function test_ping*
will be executed.