bench/interface: add arp member
Add an 'arp' member to the Interface class, to query information from
the ARP table specific to an interface. For now, this 'arp' member only
has a get method. It returns all the entries in the ARP table that are
related to this interface. An optional 'address' argument can be passed
to filter entries matching this address.