tests/helper: add StatsMonitor
Add a class to help monitoring statistics during a test. It's intended
to be used as a context, and takes snapshots of monitored interfaces'
statistics when entering and leaving the context.

When two snapshots have been taken, statistics can be accessed like
`monitor.rx_octets`, it will compute the total difference of rx_octets
on all monitored interfaces between the two snapshots.

More than two snapshots can be taken, but only the last two will be
used.