bench/control: make checkExitCode take last exit code as argument
Instead of storing the last exit code as class attribute, just pass it
to the function in charge of checking if the command succeeded or not.
As it doesn't use "self" anymore, it can be made static.

getLastExitCode is dropped as it's not used anymore. This function was
broken during the refactoring of the "execute" function.

Fixes: 8f6d730b5f "bench/control: factorize `execute` function"