bench/bench: store the dry_run state
unittest decorators are evaluated when tests are discovered and loaded.
As this happens before bench.connect() is called, tests will always see
bench.target.control.is_connected() return False.

So instead on relying on that, introduce bench.dry_run attribute. This
will be used to store information about how the user intends to run the
bench, either in dry mode or not. As this attribute is set before tests
are loaded, decorators will be able to use that determine if tests must
be skipped or not. This obsoletes bench.control.is_connected() as it was
only used (and wrongly) by unittest decorators.