bench/runner: do sys.exit at module level
This is a preparation step for the setup script. When a function is used
as an entry point with `setup.py`, the result of this function is
already wrapped in a call to sys.exit.

So let start_bench and main be regular functions that simply return a
value, and do the sys.exit at module level.