tests/helpers: add helper to wait for interfaces to be up
A common use-case is to put interfaces in the UP state, and then use
them for tests. But we might end up in a situation where the interface
is not in the operational state "up" before the test is started, so it's
possible that a test works on a board, but not on another one.