ci: use travis to automatically build documentation
Currently, HTML pages must be built manually through:
        cd docs && rake

This is not very practical and there are good chances that the doc will
soon fall out of sync. To prevent that, use Travis to build the
documentation and push it to `netdsa/netdsa.github.io`. That way, online
documentation should always be in sync with the code.